+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 - Message logging improved when a MIDI CC is received from Nymphes on an unexpected channel.
 - Fixed bug in NymphesOSC _get_local_ip_address which could cause a crash
 - Added utilities module to nymphes_osc
+- Implementing Preset Pack Feature
+  - Added ability to save all presets to a preset pack CSV file
 
 
 ## v1.0.1